Add: Category and Product models added
Category is the type of product. It has a name and description and will be used by the product model. Product is a model that represents a product. It has a name,
description, a price that must be greater than 0, but less than 100,000, and a category which is the category model.